requirements.txt
# Core dependencies
mitmproxy>=10.0.0
Flask>=3.0.0
PyYAML>=6.0.0

# Testing
pytest>=7.4.0
pytest-mock>=3.12.0
pytest-cov>=4.1.0
